% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socs.R
\name{FOCS}
\alias{FOCS}
\title{Run FOCS significance computation on a list of communities.}
\usage{
FOCS(community_list, network, p = 0.25, nrand = 100)
}
\arguments{
\item{community_list}{(list of int vectors) List of communities to compute significance of.}

\item{network}{(igraph) The network in igraph format.}

\item{p}{(double, default: 0.25) Proportion of community to use as the border.}

\item{nrand}{(integer, default: 100) Number of random choices of the p-scores.}
}
\description{
Run FOCS significance computation on a list of communities.
}
