% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{MtSigComm}
\alias{MtSigComm}
\title{Compute statistical significance of a single community.}
\usage{
MtSigComm(community_list, network, borderp = 0.25, nmcsims = 100)
}
\arguments{
\item{community_list}{(list of integers) List of communities to compute significance of.}

\item{network}{(igraph) The network in igraph format.}

\item{borderp}{(double, default: 0.25) Proportion of community to use as the border.}

\item{nmcsims}{(integer, default: 100) Number of monte carlo simulations of p-values to produce.}
}
\description{
Compute statistical significance of a single community.
}
