% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_function.R
\name{MtSigCommSingle}
\alias{MtSigCommSingle}
\title{Compute statistical significance of a single community.}
\usage{
MtSigCommSingle(group_s, G, adjG, edges, d, N, m, version = 3,
  borderp = 0.25, nmcsims = 100, return_mean = FALSE)
}
\arguments{
\item{group_s}{(integer) The community.}

\item{G}{(igraph) The network in igraph format.}

\item{adjG}{(dgCMatrix) The adjacency matrix of the network.}

\item{edges}{(integer matrix) The edgelist of the network.}

\item{d}{(integer) The degree vector of the network.}

\item{N}{(integer) The number of nodes in the network.}

\item{m}{(integer) Twice the number of edges in the network.}

\item{version}{(integer, default: 3) Version of the method to use.}

\item{borderp}{(double, default: 0.25) Proportion of community to use as the border.}

\item{nmcsims}{(integer, default: 100) Number of monte carlo simulations of p-values to produce.}

\item{return_mean}{(bool, default: FALSE) Whether to return mean or median of mc p-values.}
}
\description{
Compute statistical significance of a single community.
}
